utils/requirements.txt
anyio==4.3.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
Babel==2.14.0
beautifulsoup4==4.12.3
bleach==6.1.0
bs4==0.0.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
comm==0.2.1
cssselect==1.2.0
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.8
Django==5.0.3
executing==2.0.1
fake-useragent==2.0.3
fastjsonschema==2.19.1
filelock==3.13.3
fqdn==1.5.1
google-api-python-client==2.163.0 
gspread==6.2.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
importlib_metadata==8.5.0
ipykernel==6.29.2
ipython==8.22.1
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.3
json5==0.9.17
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter-events==0.9.0
jupyter-http-over-ws==0.0.8
jupyter-lsp==2.2.2
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.12.5
jupyter_server_terminals==0.5.2
jupyterlab==4.1.2
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.3
lxml==5.3.0
lxml_html_clean==0.4.1
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
mistune==3.0.2
nbclient==0.9.0
nbconvert==7.16.1
nbformat==5.9.2
nest-asyncio==1.6.0
notebook==7.1.0
notebook_shim==0.2.4
numpy==2.2.1
oauth2client==4.1.3
openpyxl==3.1.5 
outcome==1.3.0.post0
overrides==7.7.0
packaging==23.2
pandas==2.2.3
pandocfilters==1.5.1
parse==1.20.2
parso==0.8.3
platformdirs==4.2.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
psutil==5.9.8
pure-eval==0.2.2
pycparser==2.21
pydrive==1.3.1
pyee==11.1.1
Pygments==2.17.2
pyppeteer==2.0.0
pyquery==2.0.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-json-logger==2.0.7
pytz==2024.2
pywin32==306
pywinpty==2.0.12
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.33.0
requests==2.31.0
requests-html==0.10.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.18.0
selenium==4.27.1
Send2Trash==1.8.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
stack-data==0.6.3
tabulate==0.9.0
terminado==0.18.0
tinycss2==1.2.1
tornado==6.4
tqdm==4.67.1
traitlets==5.14.1
trio==0.28.0
trio-websocket==0.11.1
types-python-dateutil==2.8.19.20240106
typing_extensions==4.12.2
tzdata==2024.1
uri-template==1.3.0
urllib3==1.26.20
virtualenv==20.25.1
w3lib==2.2.1
wcwidth==0.2.13
webcolors==1.13
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
websockets==10.4
wsproto==1.2.0
zipp==3.21.0
